requirements.txt
CoffeeScript==1.0.10
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyExecJS==1.0.4
Werkzeug==0.9.6
arrow==0.4.4
bottle==0.12.7
facebook-sdk==0.4.0
itsdangerous==0.24
pyjade==3.0.0
python-dateutil==2.2
six==1.8.0
wsgiref==0.1.2
